+++ smarter/smarter/apps/plugin/api/v1/views.py
@@ -8,7 +8,7 @@ from urllib.parse import urljoin
 
 import yaml
 from django.core.exceptions import ValidationError
-from django.core.handlers.wsgi import WSGIRequest
+from django.core.handlers.asgi import ASGIRequest
 from django.http import HttpResponseRedirect, JsonResponse
 from django.shortcuts import redirect
 from rest_framework import status
@@ -39,7 +39,7 @@ logger = logging.getLogger(__name__)
 class PluginView(SmarterAuthenticatedAPIView):
     """Plugin view for smarter api."""
 
-    def get(self, request: WSGIRequest, plugin_id):
+    def get(self, request: ASGIRequest, plugin_id):
 
         if not waffle.switch_is_active(SmarterWaffleSwitches.ALLOW_API_GET):
             logger.error(
@@ -53,23 +53,23 @@ class PluginView(SmarterAuthenticatedAPIView):
 
         return get_plugin(request, plugin_id)
 
-    def put(self, request: WSGIRequest):
+    def put(self, request: ASGIRequest):
         return create_plugin(request)
 
-    def post(self, request: WSGIRequest):
+    def post(self, request: ASGIRequest):
         return create_plugin(request)
 
-    def patch(self, request: WSGIRequest):
+    def patch(self, request: ASGIRequest):
         return update_plugin(request)
 
-    def delete(self, request: WSGIRequest, plugin_id):
+    def delete(self, request: ASGIRequest, plugin_id):
         return delete_plugin(request, plugin_id)
 
 
 class PluginCloneView(SmarterAuthenticatedAPIView):
     """Plugin clone view for smarter api."""
 
-    def post(self, request: WSGIRequest, plugin_id, new_name):
+    def post(self, request: ASGIRequest, plugin_id, new_name):
 
         user = get_resolved_user(request.user)
         if not user:
@@ -104,7 +104,7 @@ class PluginListView(SmarterAuthenticatedListAPIView):
 class AddPluginExamplesView(SmarterAuthenticatedAPIView):
     """Add example plugins to a user profile."""
 
-    def post(self, request: WSGIRequest, user_id=None):
+    def post(self, request: ASGIRequest, user_id=None):
         @cache_results()
         def cached_user_by_id(user_id: int) -> Optional[User]:
             """Retrieve User by ID with caching."""
@@ -155,14 +155,14 @@ class PluginUploadView(SmarterAuthenticatedAPIView):
 
         raise SmarterValueError("Invalid data format: expected JSON or YAML.")
 
-    def _create(self, request: WSGIRequest):
+    def _create(self, request: ASGIRequest):
         data = self.parse_yaml_file(data=request.body.decode("utf-8"))
         return create_plugin(request=request, data=data)
 
-    def put(self, request: WSGIRequest):
+    def put(self, request: ASGIRequest):
         return self._create(request)
 
-    def post(self, request: WSGIRequest):
+    def post(self, request: ASGIRequest):
         return self._create(request)
 
 
@@ -243,7 +243,7 @@ def create_plugin(request, data: Optional[dict] = None):
     return HttpResponseRedirect(plugins_api_url + str(plugin.id) + "/")
 
 
-def update_plugin(request: WSGIRequest):
+def update_plugin(request: ASGIRequest):
     """update a plugin from a json representation in the body of the request."""
     user = get_resolved_user(request.user)
     data: str