+++ smarter/smarter/lib/journal/http.py
@@ -123,13 +123,13 @@ class SmarterJournaledJsonResponse(JsonResponse, SmarterHelperMixin):
                 return {}
 
         if waffle.switch_is_active(SmarterWaffleSwitches.ENABLE_JOURNAL):
-            # WSGIRequest can be finicky depending on the kind of response we're dealing with.
+            # ASGIRequest can be finicky depending on the kind of response we're dealing with.
             # in general, we only want the user object if it's authenticated, which happens
             # when the user is logged in to the web console, and also when the request is made
             # via api, with a valid api key.
             #
             # Original exception text was:
-            # 'WSGIRequest' object has no attribute 'user'.
+            # 'ASGIRequest' object has no attribute 'user'.
             # AttributeError: 'PreparedRequest' object has no attribute 'user'
             try:
                 if is_authenticated_request(request):